+++ packages/dumb/src/index.ts
@@ -56,8 +56,9 @@ import { SimpleStatMemoized as SimpleStat } from './components/stats/simple/simp
 import { LabeledCheckbox } from './components/controls/labeled-checkbox/labeled-checkbox'
 import { CbgMedianAnimated } from './components/trends/cbg/cbg-median-animated/cbg-median-animated'
 import { Table } from './components/settings/table'
-import CbgSliceSegment from './components/trends/cbg/cbg-slice-segment/cbg-slice-segment'
+import CbgSliceSegment from './components/trends/cbg/cbg-slice/cbg-slice-segment'
 import { NoDataLabel } from './components/trends/common/no-data-label/no-data-label'
+import { CbgSliceAnimated } from './components/trends/cbg/cbg-slice/cbg-slice-animated'
 import { BgPrefs } from './models/blood-glucose.model'
 import { XAxisLabels } from './components/trends/common/x-axis-labels/x-axis-labels'
 import { XAxisTicks } from './components/trends/common/x-axis-ticks/x-axis-ticks'
@@ -71,6 +72,7 @@ export {
   CBGMeanStat,
   CbgMedianAnimated,
   CBGPercentageBarChart,
+  CbgSliceAnimated,
   CbgSliceSegment,
   CBGStandardDeviation,
   CBGStatType,